% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{abs_byname}
\alias{abs_byname}
\title{Absolute value of matrix elements}
\usage{
abs_byname(a, mc.cores = get_mc_cores())
}
\arguments{
\item{a}{a matrix or list of matrices}

\item{mc.cores}{an integer specifying the number of cores to be used.
Default is \code{get_mc_cores()} or \code{1}. 
Try \code{mc.cores = parallel::detectCores()}.}
}
\value{
a with each element replaced by its absolute value
}
\description{
Absolute value of matrix elements
}
\examples{
library(magrittr)
abs_byname(1)
abs_byname(-1)
m <- matrix(c(-10,1,1,100), nrow = 2, dimnames = list(paste0("i", 1:2), paste0("c", 1:2))) \%>\%
  setrowtype("Industry") \%>\% setcoltype("Commodity")
abs_byname(m)
}
