% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{aggregate_byname}
\alias{aggregate_byname}
\title{Aggregate rows and columns in a matrix}
\usage{
aggregate_byname(
  a,
  aggregation_map = NULL,
  margin = c(1, 2),
  pattern_type = "exact"
)
}
\arguments{
\item{a}{a matrix or list of matrices whose rows or columns are to be aggregated}

\item{aggregation_map}{a named list of rows or columns to be aggregated (or \code{NULL}). See \code{details}.}

\item{margin}{\code{1}, \code{2}, or \code{c(1, 2)} for row aggregation, column aggregation, or both}

\item{pattern_type}{See \code{make_pattern()}.}
}
\value{
a version of \code{a} with aggregated rows and/or columns
}
\description{
Rows (\verb{margin  1}), columns (\code{margin = 2}), or both (\code{margin = c(1, 2)}, the default)
are aggregated according to \code{aggregation_map}.
}
\details{
When \code{aggregation_map} is \code{NULL} (the default),
rows (or columns or both) of same name are aggregated together.

If \code{aggregation_map} is not \code{NULL}, it must be a named list.
The name of each \code{aggregation_map} item is the name of a row or column in \code{a} that will contain the specified aggregation.
The value of each item in \code{aggregation_map} must be a vector of names of rows or columns in \code{a}.
The names in the value are aggregated and inserted into \code{a} with the name of the value.
For example \code{aggregation_map = list(new_row = c("r1", "r2"))}
will aggregate rows "r1" and "r2", delete rows "r1" and "r2", and insert a new row
whose name is "new_row" and whose value is the sum of rows "r1" and "r2'.

The items in \code{aggregation_map} are interpreted as regular expressions, and
they are escaped using \code{Hmisc::escapeRegex()} prior to use.

Note that aggregation on one margin only will sort only the aggregated margin, because
the other margin is not guaranteed to have unique names.
}
