% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{all_byname}
\alias{all_byname}
\title{Are all matrix elements \code{TRUE}?}
\usage{
all_byname(a)
}
\arguments{
\item{a}{a matrix or list of matrices}
}
\value{
\code{TRUE} if all elements of \code{a} are \code{TRUE}, \code{FALSE} otherwise
}
\description{
Tells whether all elements in matrix \code{a} are true.
}
\details{
\code{a} can be a matrix or a list of matrices.
}
\examples{
all_byname(matrix(rep(TRUE, times = 4), nrow = 2, ncol = 2))
all_byname(matrix(c(TRUE, FALSE), nrow = 2, ncol = 1))
}
