% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{apply_byname}
\alias{apply_byname}
\title{Apply an arbitrary function in the byname framework}
\usage{
apply_byname(X, MARGIN, FUN)
}
\arguments{
\item{X}{an object (usually a list or a matrix) over which \code{FUN} should be applied}

\item{MARGIN}{the margin of a matrix over which \code{FUN} should be applied. 
\code{1} for rows, \code{2} for columns, \code{1,2} for both columns and rows.}

\item{FUN}{the function to be applied to \code{X}}
}
\value{
if \code{X} is a list, a list same length as \code{X} that is the results of applying \code{FUN} to each element of \code{X}.
if \code{X} is a matrix, the result of applying \code{FUN} to that matrix on the \code{MARGIN}.
}
\description{
Allows an arbitrary function (\code{FUN}) to be applied in the byname framework.
}
\examples{
m <- matrix(c(-1:2), nrow = 2, dimnames = list(c("p1", "p2"), c("i1", "i2")))
apply_byname(m, c(1,2), function(e){ifelse(e < 0, 0, e)})
}
