% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_byname.R
\name{as.mat_byname}
\alias{as.mat_byname}
\title{Convert a \code{matrix}-like object to a \code{mat_byname}}
\usage{
as.mat_byname(x)
}
\arguments{
\item{x}{the matrix-like object to be converted to a \code{mat_byname} object}
}
\value{
the \code{mat_byname} object
}
\description{
Convert a \code{matrix}-like object to a \code{mat_byname}
}
\examples{
as.mat_byname(matrix(1:2))
}
