% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{binaryfunctionhelper_byname}
\alias{binaryfunctionhelper_byname}
\title{Helper function for binary _byname operations}
\usage{
binaryfunctionhelper_byname(FUN, OPER, a, b)
}
\arguments{
\item{FUN}{the function that is calling this helper}

\item{OPER}{the operation that \code{FUN} performs}

\item{a}{the first operand}

\item{b}{the second operand}
}
\value{
the result of applying \code{OPER} to \code{a} and \code{b} via \code{FUN}
}
\description{
Puts all tasks associated with performing a binary matrix operation _byname in one place,
including dealing intelligently with lists.
}
