% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{clean_byname}
\alias{clean_byname}
\title{Cleans (deletes) rows or columns of matrices that contain exclusively \code{clean_value}}
\usage{
clean_byname(m, margin, clean_value = 0)
}
\arguments{
\item{m}{the matrix to be cleaned}

\item{margin}{the dimension over which cleaning should occur}

\item{clean_value}{the undesirable value

When a row (when \code{margin = 1}) or a column (when \code{margin = 2}) 
contains exclusively \code{clean_value}, the row or column is deleted from the matrix.}
}
\value{
a "cleaned" matrix, expunged of rows or columns that contain exclusively \code{clean_value}.
}
\description{
Cleans (deletes) rows or columns of matrices that contain exclusively \code{clean_value}
}
\examples{
m <- matrix(c(-20, 1, -20, 2), nrow = 2, dimnames = list(c("r1", "r2"), c("c1", "c2")))
m
m \%>\% clean_byname(margin = 1, clean_value = -20) # Eliminates -20, -20 row
m \%>\% clean_byname(margin = 2) # Nothing cleaned, because no columns contain all 0's (the default clean_value).
# Also works with lists
list(m, m) \%>\% clean_byname(margin = 1, clean_value = -20)
# Also works with data frames
DF <- data.frame(m = I(list()))
DF[[1,"m"]] <- m
DF[[2,"m"]] <- m
DF \%>\% clean_byname(margin = 1, clean_value = -20)
}
