% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{colsums_byname}
\alias{colsums_byname}
\title{Column sums, sorted by name}
\usage{
colsums_byname(m, colname = NA)
}
\arguments{
\item{m}{a matrix or data frame from which column sums are desired.}

\item{colname}{name of the output column containing column sums

Calculates column sums for a matrix by pre-multiplying by an identity vector (containins all 1's).
In contrast to \code{colSums} (which returns a \code{numeric} result), 
the return value from \code{colsums_byname} is a matrix.
An optional \code{rowname} can be supplied.}
}
\value{
a row vector of type \code{matrix} containing the column sums of \code{m}.
}
\description{
Column sums, sorted by name
}
\examples{
m <- matrix(c(1:6), nrow = 2, dimnames = list(paste0("i", 1:2), paste0("c", 3:1))) \%>\%
  setrowtype("Industries") \%>\% setcoltype("Commodities")
colsums_byname(m)
colsums_byname(m, "E.ktoe")
m \%>\% colsums_byname \%>\% rowsums_byname
# This also works with lists
colsums_byname(list(m, m))
colsums_byname(list(m, m), "E.ktoe")
colsums_byname(list(m, m), NA)
colsums_byname(list(m, m), NULL)
DF <- data.frame(m = I(list()))
DF[[1,"m"]] <- m
DF[[2,"m"]] <- m
colsums_byname(DF$m[[1]])
colsums_byname(DF$m)
colsums_byname(DF$m, "sums")
res <- DF \%>\% mutate(
  cs = colsums_byname(m),
  cs2 = colsums_byname(m, rowname = "sum")
)
res$cs2
}
