% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{coltype}
\alias{coltype}
\title{Column type}
\usage{
coltype(x)
}
\arguments{
\item{x}{the object from which you want to extract column types}
}
\value{
the column type of \code{x}
}
\description{
Extracts column type of \code{x}.
}
\examples{
library(magrittr)
commoditynames <- c("c1", "c2")
industrynames <- c("i1", "i2")
U <- matrix(1:4, ncol = 2, dimnames = list(commoditynames, industrynames)) \%>\%
  setrowtype(rowtype = "Commodities") \%>\% setcoltype("Industries")
coltype(U)
# This also works for lists
coltype(list(U,U))
}
