% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{constant_vector_byname}
\alias{constant_vector_byname}
\title{Title}
\usage{
constant_vector_byname(data, k, col_name, rowtype = NULL, coltype = NULL)
}
\arguments{
\item{data}{}

\item{k}{}

\item{col_name}{}

\item{rowtype}{}

\item{coltype}{}
}
\value{

}
\description{
Title
}
