% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply.R
\name{cumapply_byname}
\alias{cumapply_byname}
\title{Cumulatively apply a function to a list of matrices or numbers}
\usage{
cumapply_byname(FUN, m)
}
\arguments{
\item{FUN}{the function to be applied}

\item{m}{the list of matrices or numbers to which \code{FUN} will be applied cumulatively}
}
\value{
a list of same length as \code{m} 
        containing the cumulative application of \code{FUN} to \code{m}
}
\description{
\code{FUN} must be a binary function that also accepts a single argument.
The result is a list with first element \code{FUN(m[[1]])}.
For \code{i >= 2}, elements are \code{FUN(m[[i]], m[[i-1]])}.
}
\examples{
cumapply_byname(sum, list(1, 2, 3, 4))
cumapply_byname(sum_byname, list(1, 2, 3, 4))
cumapply_byname(prod, list(1, 2, 3, 4))
cumapply_byname(elementproduct_byname, list(1, 2, 3, 4))
}
