% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{cumsum_byname}
\alias{cumsum_byname}
\title{Cumulative sum that respects row and column names}
\usage{
cumsum_byname(m)
}
\arguments{
\item{m}{a number, list of numbers, matrix or list of matrices for which cumulative sum is desired}
}
\value{
a single number, list of numbers, a single matrix, or a list of matrices,
        depending on the nature of \code{m}
}
\description{
Provides cumulative sums along a list or column of a data frame.
If \code{m} is a single number, \code{m} is returned.
If \code{m} is a list of numbers, a list representing the cumulative sum of the numbers is returned.
If \code{m} is a single matrix, \code{m} is returned.
If \code{m} is a list of matrices, a list representing the cumulative sum
of the matrices is returned. 
In this case, each entry in the list is sum "by name," such that row and column names 
of the matrices are respected.
}
\details{
This function respects groups.
}
\examples{

}
