% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{elementlog_byname}
\alias{elementlog_byname}
\title{Logarithm of matrix elements}
\usage{
elementlog_byname(M, base = exp(1))
}
\arguments{
\item{M}{a matrix of list of matrices}

\item{base}{the base of the logarithm (default is \code{exp(1)}, giving the natural logarithm)}
}
\value{
M with each element replaced by its base \code{base} logarithm
}
\description{
Specify the base of the log with \code{base} argument.
}
\examples{
library(magrittr)
elementlog_byname(exp(1))
m <- matrix(c(10,1,1,100), nrow = 2, dimnames = list(paste0("i", 1:2), paste0("c", 1:2))) \%>\%
  setrowtype("Industry") \%>\% setcoltype("Commodity")
elementlog_byname(m)
elementlog_byname(m, base = 10)
}
