% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary.R
\name{elementpow_byname}
\alias{elementpow_byname}
\title{Powers of matrix elements}
\usage{
elementpow_byname(M, pow)
}
\arguments{
\item{M}{a matrix of list of matrices}

\item{pow}{the power to which elements of \code{M} will be raised}
}
\value{
\code{M} with each element raised to \code{pow}
}
\description{
Gives the result of raising all elements of a matrix or list of matrices to a power.
}
\examples{
library(magrittr)
elementpow_byname(1)
m <- matrix(c(log(10),log(1),log(1),log(100)), 
  nrow = 2, dimnames = list(paste0("i", 1:2), paste0("c", 1:2))) \%>\%
  setrowtype("Industry") \%>\% setcoltype("Commodity")
elementexp_byname(m)
}
