% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{exp_byname}
\alias{exp_byname}
\title{Exponential of matrix elements}
\usage{
exp_byname(a)
}
\arguments{
\item{a}{a matrix of list of matrices}
}
\value{
\code{M} with each element replaced by its exponential
}
\description{
Gives the exponential of all elements of a matrix or list of matrices
}
\examples{
exp_byname(1)
m <- matrix(c(log(10),log(1),
              log(1),log(100)), 
              byrow = TRUE, nrow = 2, ncol = 2,
              dimnames = list(paste0("i", 1:2), paste0("c", 1:2))) \%>\%
  setrowtype("Industry") \%>\% setcoltype("Commodity")
m
exp_byname(m)
}
