% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{get_mc_cores}
\alias{get_mc_cores}
\title{Get number of cores to be used for parallelized matsbyname functions}
\usage{
get_mc_cores()
}
\value{
the package-wide value for number of cores to be used for \pkg{matsbyname} functions
}
\description{
Get the package-wide value for the number of cores to be used for parallized
\pkg{matsbyname} functions.
The default value is \code{1}.
}
\details{
If the \code{mc.cores} argument is set on a \pkg{matsbyname} function, 
the argument's value is used instead of the package-wide default value obtained from this function.
}
\examples{
get_mc_cores()
}
