% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_triplet.R
\name{get_row_col_index_maps}
\alias{get_row_col_index_maps}
\title{Figure out row and column index maps}
\usage{
get_row_col_index_maps(a_mat, ind_maps)
}
\arguments{
\item{a_mat}{A matrix for which index maps should be determined.}

\item{ind_maps}{A list of two or more data frames
of potential index maps.}
}
\value{
A list of two data frames.
The first data frame is the index map for the rows of \code{a_mat}.
The second data frame is the index map for the columns of \code{a_mat}.
}
\description{
The \code{index_map} argument can take several forms.
This function figures out (for a given \code{a_mat})
the index maps for rows (first data frame in the return list)
and columns (second data frame in the return list).
}
\details{
This is a non-exported function meant only for internal use.
}
