% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{Iminus_byname}
\alias{Iminus_byname}
\title{Subtract a matrix with named rows and columns from a suitably named and sized identity matrix (\code{I})}
\usage{
Iminus_byname(m)
}
\arguments{
\item{m}{the matrix to be subtracted from \code{I}}
}
\value{
The difference between an identity matrix (\code{I}) and \code{m} 
(whose rows and columns have been completed and sorted)
}
\description{
\code{m} must be square.
Note: the order of rows and columns of \code{m} may before subtracting from \code{I}, 
because the rows and columns are sorted by name prior to subtracting from \code{I}.
}
\examples{
m <- matrix(c(-21, -12, -21, -10), ncol = 2, dimnames = list(c("b", "a"), c("b", "a"))) \%>\%
  setrowtype("Industries") \%>\% setcoltype("Commodities")
diag(1, nrow = 2) - m # Rows and columns of A are unsorted
Iminus_byname(m) # Rows and columns of A are sorted prior to subtracting from the identity matrix
# This also works with lists
Iminus_byname(list(m,m))
}
