% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{invert_byname}
\alias{invert_byname}
\title{Invert a matrix}
\usage{
invert_byname(a, mc.cores = get_mc_cores())
}
\arguments{
\item{a}{the matrix to be inverted. \code{m} must be square.}

\item{mc.cores}{an integer specifying the number of cores to be used.
Default is \code{get_mc_cores()} or \code{1}. 
Try \code{mc.cores = parallel::detectCores()}.}
}
\value{
the inversion
}
\description{
This function transposes row and column names as well as row and column types.
Rows and columns of \code{m} are sorted prior to inverting.
}
\examples{
library(magrittr)
m <- matrix(c(10,0,0,100), nrow = 2, dimnames = list(paste0("i", 1:2), paste0("c", 1:2))) \%>\%
  setrowtype("Industry") \%>\% setcoltype("Commodity")
invert_byname(m)
matrixproduct_byname(m, invert_byname(m))
matrixproduct_byname(invert_byname(m), m)
invert_byname(list(m,m))
}
