% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_byname.R
\name{is.mat_byname}
\alias{is.mat_byname}
\title{Test if an argument is a \code{mat_byname}}
\usage{
is.mat_byname(x)
}
\arguments{
\item{x}{the object to be tested}
}
\value{
\code{TRUE} if \code{x} is a \code{mat_byname}, \code{FALSE} otherwise
}
\description{
Test if an argument is a \code{mat_byname}
}
\examples{
is.mat_byname(matrix(1:2))
is.mat_byname(mat_byname(1:2))
}
