% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Matrix.R
\name{is.Matrix}
\alias{is.Matrix}
\title{Is an object a Matrix?}
\usage{
is.Matrix(a)
}
\arguments{
\item{a}{The object to be queried if it is Matrix.}
}
\value{
A boolean. \code{TRUE} if \code{a} is a \code{Matrix}, \code{FALSE} otherwise.
}
\description{
Arguably, this function should be in the \code{Matrix} package,
but it is not.
We include it here for convenience.
}
\details{
This function is not vectorized.

\code{is.Matrix()} is a wrapper for \verb{inherits(a, "Matrix)}.
}
\examples{
is.Matrix(matrix(42))
is.Matrix(Matrix::Matrix(42))
}
