% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_Matrix.R
\name{is_matrix_or_Matrix}
\alias{is_matrix_or_Matrix}
\title{Tells whether an object is one of a matrix or a Matrix}
\usage{
is_matrix_or_Matrix(a)
}
\arguments{
\item{a}{The object about which we want to know if it is a \code{matrix} or a \code{Matrix}.}
}
\value{
\code{TRUE} when \code{a} is a \code{matrix} or a \code{Matrix}. \code{FALSE} otherwise.
}
\description{
Often, it helps to know whether an object is a \code{matrix} or a \code{Matrix},
and you don't care which.
This function helps in those situations.
}
\examples{
is_matrix_or_Matrix(42)
is_matrix_or_Matrix(matrix(42))
is_matrix_or_Matrix(Matrix::Matrix(42))
is_matrix_or_Matrix(matsbyname::Matrix(42))
}
