% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{list_of_rows_or_cols}
\alias{list_of_rows_or_cols}
\title{Named list of rows or columns of matrices}
\usage{
list_of_rows_or_cols(m, margin)
}
\arguments{
\item{m}{a matrix or list of matrices (say, from a column in a data frame)}

\item{margin}{the dimension of the matrices to be extracted (1 = rows, 2 = columns)}
}
\value{
a named list of rows or columns extracted from \code{m}
}
\description{
Named list of rows or columns of matrices
}
