% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{logarithmicmean_byname}
\alias{logarithmicmean_byname}
\title{Name- and element-wise logarithmic mean of matrices.}
\usage{
logarithmicmean_byname(X1, X2, base = exp(1))
}
\arguments{
\item{X1}{first operand (a matrix or constant value or lists of same).}

\item{X2}{second operand (a matrix or constant value or lists of same).}

\item{base}{the base of the logarithm used when computing the logarithmic mean.
(Default is \code{base = exp(1)}.)}
}
\value{
A matrix representing the name-wise logarithmic mean 
        of \strong{\code{X1}} and \strong{\code{X2}}.
}
\description{
The logarithmic mean of corresponding entries of \strong{\code{X1}} and \strong{\code{X2}} is 
\code{0} if \code{x1 = 0} or \code{x2 = 0}, 
\code{x1} if \code{x1 = x2}, or
\code{(x2 - x1) / (log(x2) - log(x1))} otherwise.
}
\details{
This function performs a union and sorting of row and column names 
prior to performing logarithmic mean.
Zeroes are inserted for missing matrix elements.

Internally, the third condition is implemented as 
\code{(y - x) / log(y/x)}.

Note that \code{(x2 - x1) / log(x2/x1) = (x1 - x2) / log(x1/x2)},
so logarithmic mean is commutative;
the order of arguments \strong{\code{X1}} and \strong{\code{X2}}
does not change the result.
}
\examples{
library(magrittr)
library(dplyr)
}
