% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompletingMatrices.R
\name{make_list}
\alias{make_list}
\title{Makes a list of items in x, regardless of x's type}
\usage{
make_list(x, n, lenx = ifelse(is.vector(x), length(x), 1))
}
\arguments{
\item{x}{the object to be duplicated}

\item{n}{the number of times to be duplicated}

\item{lenx}{the length of item \code{x}. Normally \code{lenx} is taken from \code{length(x)},
but if \code{x} is itself a \code{list}, you may wish for the \code{list} to be duplicated several
times. In that case, set \code{lenx = 1}.}
}
\value{
a list of \code{x} duplicated \code{n} times
}
\description{
Repeats \code{x} as necessary to make \code{n} of them.
Does not try to simplify \code{x}.
}
