% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{make_pattern}
\alias{make_pattern}
\title{Create regex patterns for row and column name selection}
\usage{
make_pattern(row_col_names, pattern_type = c("exact", "leading", "trailing",
  "anywhere"))
}
\arguments{
\item{row_col_names}{a vector of row and column names}

\item{pattern_type}{one of \code{exact}, \code{leading}, \code{trailing}, or \code{anywhere}.
\code{exact} produces a pattern that selects row or column names by exact match.
\code{leading} produces a pattern that selectes row or column names if the item in \code{row_col_names} matches
the beginning of the row or column name.
\code{trailing} produces a pattern that selectes row or column names if the item in \code{row_col_names} matches
the end of the row or column name.
\code{anywhere} produces a pattern that selectes row or column names if the item in \code{row_col_names} matches
any substring of the row or column name.}
}
\value{
an extended regex pattern suitable for use with \code{select_rows_byname} or \code{select_cols_byname}.
}
\description{
Create regex patterns for row and column name selection
}
\examples{
make_pattern(row_col_names = c("a", "b"), pattern_type = "exact")
}
