% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_byname.R
\name{mat_byname}
\alias{mat_byname}
\title{Create a \code{mat_byname} object}
\usage{
mat_byname(data = NA, nrow = 1, ncol = 1, byrow = FALSE,
  dimnames = NULL, m = matrix(data = data, nrow = nrow, ncol = ncol, byrow =
  byrow, dimnames = dimnames), rowtype = NULL, coltype = NULL)
}
\arguments{
\item{data}{a vector of values for \code{m}}

\item{nrow}{the number of rows in \code{m}}

\item{ncol}{the number of columns in \code{m}}

\item{byrow}{tells whether to fill \code{m} by rows (\code{TRUE}) or columns (\code{FALSE})}

\item{dimnames}{row and column names for \code{m}}

\item{m}{a \code{matrix} object}

\item{rowtype}{the type for rows}

\item{coltype}{the type for columns}
}
\value{
a \code{mat_byname} object
}
\description{
\code{m} should be a matrix.
}
\details{
If \code{is.null(m)}, \code{NULL} is returned.
If \code{all(is.na(m))}, \code{NA} is returned.
If \code{m} is not a \code{matrix}, an error is given.
}
\examples{
mat_byname(matrix(c(1:2)))
}
