% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_vecs.R
\name{mat_from_store_byname}
\alias{mat_from_store_byname}
\title{Create a matrix (or vector) with labels from a matrix and values from a "store"}
\usage{
mat_from_store_byname(
  a,
  v,
  a_piece = "all",
  v_piece = "all",
  margin = 1,
  margin_v = 1,
  notation = if (is.list(a)) {
     list(RCLabels::bracket_notation)
 } else {
    
    RCLabels::bracket_notation
 },
  prepositions = if (is.list(a)) {
     list(RCLabels::prepositions_list)
 } else {
    
    RCLabels::prepositions_list
 },
  missing = NA_real_
)
}
\arguments{
\item{a}{A matrix from which row or column labels are taken.
Can also be a list or the name of a column in a data frame.}

\item{v}{A matrix or vector from which values are taken, when \code{a_piece}
of names on \code{a}'s \code{margin} matches \code{v_piece} of names on \code{v}'s \code{margin_v}.
Can also be a list or the name of a column in a data frame.}

\item{a_piece}{The piece of labels on \code{a}'s \code{margin} that is to be matched. Default is "all".}

\item{v_piece}{The piece of labels on \code{v}'s \code{margin_v} that is to be matched. Default is "all".}

\item{margin}{Tells whether to assess the rows (\code{1}) or columns (\code{2}) of \code{a}
when creating the outgoing vector.
Default is \code{1}.}

\item{margin_v}{Tells whether to assess the rows (\code{1}) or columns (\code{2}) of \code{v}
when creating the outgoing vector.
Default is \code{1}.}

\item{notation}{The notation for the row and column labels.
Default is \code{RCLabels::bracket_notation}, wrapped as a list if \code{a} is a list.}

\item{prepositions}{The strings that will count for prepositions.
Default is \code{RCLabels::prepositions}, wrapped as a list if \code{a} is a list.}

\item{missing}{The value used when the desired value is not found in \code{v}.
Default is \code{NA_real_}.}
}
\value{
A column vector with names from \code{a} and values from \code{v}.
}
\description{
When a matrix (\code{a}) is multiplied by another matrix or vector (\code{v}) byname,
naming can be tricky.
There are times when pieces of the labels in \code{v} should be matched to
pieces of the \code{a} labels.
This function helps by performing the matching byname.
For this function, matrix or vector \code{v} is considered a store
of values, rows, or columns
from which the output vector or matrix is constructed
using special matching rules between matrix \code{a} and matrix or vector \code{v}.
}
\details{
The output of this function is a matrix or vector.
The names of the output rows are taken from
the \code{margin} of \code{a}.
(\code{a = 1} means rows; \code{a = 2} means columns.)
The names of the other dimension of the output are taken from \code{v},
but not \code{margin}.
Rather, the names of the other dimension of the output are taken from
the other dimension of \code{v}.
If \code{margin_v} is \code{1}, the names of the other dimension of the output
are taken from columns of \code{v}.
If \code{margin_v} is \code{2}, the names of the other dimension of the output
are taken from rows of \code{v}.

The values of the output matrix are obtained from \code{v}
when \code{a_piece} matches \code{v_piece} using the \code{RCLabels} package.
The \code{v_piece}s of \code{v}'s \code{margin_v} must be unique.
The default values for \code{a_piece} and \code{v_piece} are "all",
meaning that the entire label of \code{a} should be matched to the entire
label of \code{v}.
Other options for \code{a_piece} and \code{v_piece} are "pref" and "suff",
which will match the prefix or suffix of the labels.
Alternatively, prepositions can be given such that
objects of prepositions will be matched.
Examples include "from" or "in".
Row and column types from \code{v} are applied to the output.
If the piece given in \code{a_piece} is not present in row or column names of \code{a},
\code{NA_real_} is returned.
If the piece given in \code{v_piece} is not present in row or column names of \code{v},
\code{NA_real_} is returned.

Note that \code{notation} and \code{prepositions} should be lists if \code{a} is a list
but a single value otherwise.
The default values of \code{notation} and \code{prepositions} take care of this requirement,
switching on the type of \code{a} (list or not).

The class of the output object is determined from \code{a}.
If \code{a} is a \code{Matrix}, the output will be a \code{Matrix}.
Otherwise, the output will be a \code{matrix}.

\code{mat_from_store_byname()} is a more-capable replacement for
the deprecated \code{vec_from_store_byname()}.
\code{mat_from_store_byname()} allows matrices in the \code{v} argument.
If \code{vec_from_store_byname()} was used previously with a column vector
in the \code{v} argument and a column vector was the desired output,
\code{mat_from_store_byname()} should be a drop-in replacement.
}
\examples{
a <- matrix(42, nrow = 3, ncol = 5, 
            dimnames = list(c("Electricity [from b in c]", 
                              "Coal [from e in f]", 
                              "Crude oil [from Production in USA]"), 
                            c("Main activity producer electricity plants", 
                              "Wind turbines", 
                              "Oil refineries", 
                              "Coal mines", 
                              "Automobiles"))) \%>\%
  setrowtype("Product") \%>\% setcoltype("Industry")
a
v <- matrix(1:7, nrow = 7, ncol = 1, 
            dimnames = list(c("Electricity", 
                              "Peat", 
                              "Hydro", 
                              "Crude oil",
                              "Coal", 
                              "Hard coal (if no detail)", 
                              "Brown coal"), 
                            "phi")) \%>\%
  setrowtype("Product") \%>\% setcoltype("phi")
v
vec_from_store_byname(a, v, a_piece = "pref")
vec_from_store_byname(a, v, a_piece = "noun")

v2 <- matrix(1:7, nrow = 7, ncol = 1, 
             dimnames = list(c("Electricity", 
                               "Peat", 
                               "USA", 
                               "c",
                               "Coal", 
                               "Hard coal (if no detail)", 
                               "f"), 
                             "phi")) \%>\%
  setrowtype("Product") \%>\% setcoltype("phi")
vec_from_store_byname(a, v2, a_piece = "in")

# Works with lists
v3 <- matrix(1:7, nrow = 7, ncol = 1, 
             dimnames = list(c("Electricity [from USA]", 
                               "Peat [from nowhere]", 
                               "Production [from GHA]", 
                               "e [from ZAF]",
                               "Coal [from AUS]", 
                               "Hard coal (if no detail) [from GBR]", 
                               "b [from Nebraska]"), 
                             "phi")) \%>\%
  setrowtype("Product") \%>\% setcoltype("phi")
a_list <- list(a, a)
v_list <- list(v3, v3)
vec_from_store_byname(a_list, v_list, a_piece = "in", v_piece = "from")

# Also works in a data frame
df <- tibble::tibble(a = list(a, a, a), 
                     v = list(v3, v3, v3))
df \%>\%
  dplyr::mutate(
    actual = vec_from_store_byname(a = a, v = v, a_piece = "in", v_piece = "from")
  )
}
