% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{matricize_byname}
\alias{matricize_byname}
\title{Matricize a vector}
\usage{
matricize_byname(a, notation_list)
}
\arguments{
\item{a}{a row (column) vector to be converted to a matrix based on its row (column) names.}

\item{notation_list}{a string vector created by \code{row_col_notation_list()} that identifies the notation for row and column names.}
}
\value{
a matrix created from vector \code{a}.
}
\description{
Matricize a vector
}
\examples{
v <- matrix(c(1,
              2,
              3, 
              4), 
            nrow = 4, ncol = 1, dimnames = list(c("p1 -> i1", 
                                                  "p2 -> i1", 
                                                  "p1 -> i2", 
                                                  "p2 -> i2"))) \%>\% 
  setrowtype("Products -> Industries")
# Default separator is " -> ".
matricize_byname(v, notation_list = arrow_notation_list())
}
