% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matsbyname-package.R
\docType{package}
\name{matsbyname-package}
\alias{matsbyname}
\alias{matsbyname-package}
\title{matsbyname: An Implementation of Matrix Mathematics that Respects Row and Column Names}
\description{
An implementation of matrix mathematics wherein operations are performed "by name."
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MatthewHeun/matsbyname}
  \item \url{https://matthewheun.github.io/matsbyname/}
  \item Report bugs at \url{https://github.com/MatthewHeun/matsbyname/issues}
}

}
\author{
\strong{Maintainer}: Matthew Heun \email{matthew.heun@me.com} (\href{https://orcid.org/0000-0002-7438-214X}{ORCID})

Authors:
\itemize{
  \item Matthew Heun \email{matthew.heun@me.com} (\href{https://orcid.org/0000-0002-7438-214X}{ORCID})
}

}
\keyword{internal}
