% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{mcmutate}
\alias{mcmutate}
\title{Multi-core mutate}
\usage{
mcmutate(.data, mc.cores = get_mc_cores(), ...,
  mc.args = list(mc.preschedule = FALSE, mc.set.seed = TRUE, mc.silent =
  FALSE, mc.cores = mc.cores, mc.cleanup = TRUE, mc.allow.recursive = TRUE))
}
\arguments{
\item{.data}{a data frame to mutate using multiple cores}

\item{mc.cores}{the number of sub-data.frames created}

\item{...}{name-value pairs of expressions passed to \code{\link[dplyr]{mutate}}}

\item{mc.args}{arguments passed to \code{\link[parallel]{mclapply}}}
}
\value{
the mutated data frame
}
\description{
Breaks the data frame into equal-length subsets of rows,
mutates each one, and recombines.
}
\details{
Arguments to \code{\link{mclapply}} are provided by the \code{mc.cores} argument.
Note that \code{mc.preschedule} defaults to \code{FALSE},
because this function does its own prescheduling.
}
\examples{
data.frame(a = c(1,2,3), b = c(4,5,6)) \%>\% mcmutate(mc.cores = 2L, c = a + b)
}
