% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ncol_byname}
\alias{ncol_byname}
\title{Get the number of columns in a "byname" matrix.}
\usage{
ncol_byname(a)
}
\arguments{
\item{a}{A matrix or a column of a data frame populated with "byname" matrices.}
}
\value{
The number of columns of the matrix, or a list containing the number of columns
in each of the matrices contained in the column of a data frame.
}
\description{
Get the number of columns in a "byname" matrix.
}
\examples{
  dfUs <- data.frame(
year = numeric(),
matrix_byname = I(list())
)
dfUs[[1, "matrix_byname"]] <- U
dfUs[[2, "matrix_byname"]] <- U2
dfUs[[3, "matrix_byname"]] <- U3
dfUs[[1, "year"]] <- 2000
dfUs[[2, "year"]] <- 2001
dfUs[[3, "year"]] <- 2002
number_cols <- ncol_byname(dfUs$matrix_byname) \%>\% 
print()
}
