% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{nrow_byname}
\alias{nrow_byname}
\title{Get the number of rows in a "byname" matrix.}
\usage{
nrow_byname(a)
}
\arguments{
\item{a}{A matrix or a column of}
}
\value{

}
\description{
Get the number of rows in a "byname" matrix.
}
