% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{organize_args}
\alias{organize_args}
\title{Organize binary arguments}
\usage{
organize_args(a, b)
}
\arguments{
\item{a}{the first argument to be organized}

\item{b}{the second argument to be organized}
}
\value{
a list with two elements (named \code{a} and \code{b}) containing organized versions of the arguments
}
\description{
Organizes arguments of binary (2 arguments) \code{_byname} functions.
Actions performed are:
\itemize{
 \item{if only one argument is a list, make the other argument also a list of equal length.}
 \item{if both arguments are lists, ensure that they are same length.}
 \item{if one argument is a matrix and the other is a constant, make the constant into a matrix.}
 \item{ensures that row and column types match.}
 \item{completes and sorts the matrices.}
}
}
