% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply.R
\name{prepare_.FUNdots}
\alias{prepare_.FUNdots}
\title{Prepare the \code{.FUNdots} argument for \verb{*apply_byname} functions.}
\usage{
prepare_.FUNdots(a, .FUNdots)
}
\arguments{
\item{a}{the main argument to an \verb{*apply_byname} function.}

\item{.FUNdots}{additional arguments to be applied to \code{FUN} in one of the \verb{*apply_byname} functions.}
}
\value{
a reconfigured version of \code{.FUNdots}, ready for use by an \verb{*apply_byname} function.
}
\description{
This is a helper function for the various \verb{*apply_byname} functions.
}
\details{
This function looks at several cases.
\itemize{
\item If \code{.FUNdots} is not a \code{list}, this function attempts to recycle \code{.FUNdots} for use with \verb{*apply_byname}.
\item If \code{.FUNdots} is a \code{list} in which each top-level element (argument to \code{FUN}) has length 1,
this function recycles each element of \code{.FUNdots} for use across \code{a}.
\item If \code{.FUNdots} is a \code{list} in which each top-level element (argument to \code{FUN}) has same length as \code{a},
this function reorders \code{.FUNdots} for use across \code{a}.
}
}
