% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribute.R
\name{redistribute_byname}
\alias{redistribute_byname}
\title{Redistribute values from one row or column to another}
\usage{
redistribute_byname(
  a,
  rowcolnames = NULL,
  margin = c(1, 2),
  piece = "all",
  pattern_type = "exact",
  prepositions = RCLabels::prepositions_list,
  notation = RCLabels::notations_list
)
}
\arguments{
\item{a}{A matrix or a list of matrices.}

\item{rowcolnames}{The names of the rows or columns to be redistributed.}

\item{margin}{The margin of the matrix on which the \code{rowcolnames} are located.
Default is \code{c(1, 2)}, meaning that both rows (\code{1}) and columns (\code{2})
will be checked for \code{rowcolnames} and redistributed.}

\item{pattern_type}{The pattern type desired. Default is "exact".
Other options are "leading", "trailing", "anywhere",
and "literal".
See \code{\link[RCLabels:make_or_pattern]{RCLabels::make_or_pattern()}} for details.}

\item{prepositions}{Prepositions used by \code{\link[=select_rowcol_piece_byname]{select_rowcol_piece_byname()}}
for row and column name matching.
Default is \link[RCLabels:prepositions_list]{RCLabels::prepositions_list}.}

\item{notation}{The row or column notation used by \code{\link[=select_rowcol_piece_byname]{select_rowcol_piece_byname()}}
for row and column name matching.
Default is \link[RCLabels:notations_list]{RCLabels::notations_list}.}
}
\value{
A modified version of \code{a} with \code{rowcolnames} redistributed.
}
\description{
There are situations where it is helpful to
redistribute values from one row or column to another,
in proportion to remaining values in corresponding columns or rows.
This function performs the redistribution operation.
}
