% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{rename_to_pref_suff_byname}
\alias{rename_to_pref_suff_byname}
\title{Rename matrix rows and columns by prefix and suffix}
\usage{
rename_to_pref_suff_byname(a, sep, keep, margin = c(1, 2))
}
\arguments{
\item{a}{a matrix or list of matrices whose rows or columns will be renamed}

\item{sep}{a string that identifies the separator between prefix and suffix}

\item{keep}{one of "prefix" or "suffix" indicating which part of the row or column name to retain}

\item{margin}{one of \code{1}, \code{2}, or \code{c(1, 2)} where \code{1} indicates rows and \code{2} indicates columns}
}
\value{
\code{a} with potentially different row or column names
}
\description{
It can be convenient to rename rows or columns of matrices
based on retaining prefixes \emph{before} a separator or suffixes \emph{after} a separator.
This functions provides that capability.
}
\details{
Note that the \code{keep} parameter tells which to retain,
prefixes \emph{before} a separator or suffixes \emph{after} a separator.

When \code{keep} is "prefix", all characters to the right of the first (leftmost) separator and the separator itself are dropped from the names.
When \code{keep} is "suffix", all characters to the left of the last (rightmost) separator and the separator itself are dropped from the names.

If \code{sep} is not found in a name, that name is unchanged.
}
