% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{replaceNaN_byname}
\alias{replaceNaN_byname}
\title{Replace \code{NaN} values with a value}
\usage{
replaceNaN_byname(a, val = 0)
}
\arguments{
\item{a}{A matrix of list of matrices in which \code{NaN} will be replaced by \code{val}.}

\item{val}{\code{NaN}s are replace by \code{val.}}
}
\value{
A matrix or list of matrices in which all \code{NaN} are replaced by \code{val}.
}
\description{
In a matrix or within matrices in a list,
replace all \code{NaN} matrix values with \code{val.}
}
\examples{
suppressWarnings(a <- matrix(c(1, sqrt(-1))))
replaceNaN_byname(a)
replaceNaN_byname(a, 42)
}
