% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{replaceNaNWith0}
\alias{replaceNaNWith0}
\title{Replace NaN values with 0}
\usage{
replaceNaNWith0(a, mc.cores = get_mc_cores())
}
\arguments{
\item{a}{a matrix of list of matrices in which \code{NaN} will be replaced by \code{0}}

\item{mc.cores}{the number of cores to use for this calculation}
}
\value{
a matrix or list of matrices in which all \code{NaN} are replaced by \code{0}
}
\description{
In a matrix or within matrices in a list, 
replace all \code{NaN} matrix values with \code{0}.
}
\examples{
suppressWarnings(a <- matrix(c(1, sqrt(-1))))
replaceNaNWith0(a)
}
