% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{rowtype}
\alias{rowtype}
\title{Row type}
\usage{
rowtype(a, mc.cores = get_mc.cores())
}
\arguments{
\item{a}{the object from which you want to extract row types}

\item{mc.cores}{an integer specifying the number of cores to be used.
Default is \code{get_mc.cores()} or \code{1}. 
Try \code{mc.cores = parallel::detectCores()}.}
}
\value{
the row type of \code{a}
}
\description{
Extracts row type of \code{a}.
}
\examples{
library(magrittr)
library(dplyr)
commoditynames <- c("c1", "c2")
industrynames <- c("i1", "i2")
U <- matrix(1:4, ncol = 2, dimnames = list(commoditynames, industrynames)) \%>\%
  setrowtype(rowtype = "Commodities") \%>\% setcoltype("Industries")
rowtype(U)
# This also works for lists
rowtype(list(U,U))
}
