% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary.R
\name{same_structure_byname}
\alias{same_structure_byname}
\title{Test whether two matrices or lists of matrices have same structure}
\usage{
same_structure_byname(X1, X2)
}
\arguments{
\item{X1}{the first matrix to be tested}

\item{X2}{the second matrix to be tested}
}
\value{
\code{TRUE} if \code{X1} and \code{X2} have the same structure, \code{FALSE} otherwise.
}
\description{
Two matrices are said to have the same structure 
if row and column types are identical 
and 
if row and column names are identical.
}
\examples{
library(magrittr)
same_structure_byname(2, 2)
productnames <- c("p1", "p2")
industrynames <- c("i1", "i2")
U <- matrix(1:4, ncol = 2, dimnames = list(productnames, industrynames)) \%>\%
  setrowtype("Products") \%>\% setcoltype("Industries")
same_structure_byname(U, U)
same_structure_byname(U, U \%>\% setrowtype("row"))
same_structure_byname(U \%>\% setcoltype("col"), U)
# Also works with lists
same_structure_byname(list(U, U))
}
