% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{select_cols_byname}
\alias{select_cols_byname}
\title{Select columns of a matrix (or list of matrices) by name}
\usage{
select_cols_byname(m, col_names, exact = TRUE)
}
\arguments{
\item{m}{a matrix or a list of matrices}

\item{col_names}{a vector of Strings containing the column names to keep
or delete (if names are preceded by a \code{-}).}

\item{exact}{If \code{TRUE}, an exact match of column names is required.
If \code{FALSE}, \code{startsWith} will determine matches.
Retaining columns takes precedence over removing columns.
If \code{m} contains none of the requested columns to be retained, \code{NULL} is returned.
If \code{m} contains none of the requested columns to be removed, \code{m} is returned.}
}
\value{
a matrix that is a subset of \code{m} with columns selected according to \code{col_names}.
}
\description{
Select columns of a matrix (or list of matrices) by name
}
\examples{
m <- matrix(1:16, ncol = 4, dimnames=list(c(paste0("i", 1:4)), paste0("c", 1:4))) \%>\%
  setrowtype("Industries") \%>\% setcoltype("Commodities")
select_cols_byname(m, "c2")
select_cols_byname(m, c("c1", "c4"))
select_cols_byname(m, c("-c1", "-c4", "-----c3", "c2", "c4", "c4")) # Retain take precedence over remove.
select_cols_byname(m, "x") # Matches nothing. NULL is returned
select_cols_byname(m, "-x") # Matches nothing. m is returned
# Also works for lists
select_cols_byname(list(m,m), col_names = list(c("c1", "c4"), c("c2", "c3")))
select_cols_byname(list(m,m), col_names = c("c1", "c4"))
# Test inexact matches
n <- setcolnames_byname(m, c("d1", "d2", "e1", "e2"))
select_cols_byname(n, "d", exact = FALSE)
select_cols_byname(n, "-d", exact = FALSE)
}
