% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{select_zero_cols_byname}
\alias{select_zero_cols_byname}
\title{Select zero columns}
\usage{
select_zero_cols_byname(a, tol = 1e-06)
}
\arguments{
\item{a}{A matrix or a list of matrices.}

\item{tol}{The allowable deviation from 0 for any element.}
}
\value{
\code{a} with only zero columns selected.
}
\description{
Matrices with columns containing all zeroes are not invertible (singular).
To diagnose this problem, it is useful to find the zero columns
of a singular matrix.
This function selects (extracts) only the zero columns of a matrix.
}
\details{
A column is said to be a zero column if all elements are within \code{tol} of zero.
}
