% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{set_mc.cores}
\alias{set_mc.cores}
\title{Set number of cores to be used for matsbyname functions}
\usage{
set_mc.cores(mc.cores)
}
\arguments{
\item{mc.cores}{the number of cores to be set as the package-wide default}
}
\value{
nothing; this function is called for its side effects
}
\description{
Set a package-wide variable for the number of cores to be used for 
\pkg{matsbyname} functions.
The default value is \code{1}.
}
\details{
Note that all \pkg{matsbyname} functions also have a \code{mc.cores} argument.
If set, the \code{mc.cores} argument overrides the package-wide value.
}
\examples{
set_mc.cores(1)
}
