% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{setrownames_byname}
\alias{setrownames_byname}
\title{Sets row names}
\usage{
setrownames_byname(m, rownames)
}
\arguments{
\item{m}{The matrix or data frame from which row names are to be set}

\item{rownames}{The new row names}
}
\value{
a copy of \code{m} with new row names
}
\description{
Sets row names in a way that is amenable to use in chaining operations in a functional programming way
}
\examples{
m <- matrix(c(1:6), nrow = 2, dimnames = list(paste0("i", 1:2), paste0("c", 1:3))) \%>\%
  setrowtype("Industries") \%>\% setcoltype("Commodities")
setrownames_byname(m, c("a", "b"))
setrownames_byname(m \%>\% setrowtype("Industries") \%>\% setcoltype("Commodities"), c("c", "d"))
m \%>\% setrownames_byname(NULL)
m \%>\% setrownames_byname(NA)
# This also works for lists
setrownames_byname(list(m,m), list(c("a", "b"), c("c", "d")))
DF <- data.frame(m = I(list()))
DF[[1,"m"]] <- m
DF[[2,"m"]] <- m
setrownames_byname(DF$m, list(c("r1", "r2")))
setrownames_byname(DF$m, list(c("a", "b"), c("c", "d")))
DF <- DF \%>\% mutate(m = setrownames_byname(m, list(c("r1", "r2"))))
DF$m[[1]]
}
