% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_triplet.R
\name{structure_index_map}
\alias{structure_index_map}
\title{Set the structure of an index map}
\usage{
structure_index_map(index_map)
}
\arguments{
\item{index_map}{The index map data frame to be structured.}
}
\value{
A data frame with an integer column as the first column
and a character string column as the second column.
}
\description{
Index maps must be a data frame with one integer column
and one character string column.
This function verifies the structure and ensures
that the integer column is first.
}
\details{
This is a non-exported function meant only for internal use.
}
