% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_triplet.R
\name{to_named_triplet}
\alias{to_named_triplet}
\alias{to_triplet}
\alias{to_named}
\title{Convert a matrix or list of matrices between named form and indexed form}
\usage{
to_triplet(
  a,
  index_map,
  row_index_colname = "i",
  col_index_colname = "j",
  val_colname = "x",
  rownames_colname = "rownames",
  colnames_colname = "colnames"
)

to_named(
  a,
  index_map,
  matrix_class = c("matrix", "Matrix"),
  row_index_colname = "i",
  col_index_colname = "j",
  val_colname = "x"
)
}
\arguments{
\item{a}{For \code{\link[=to_triplet]{to_triplet()}}, a matrix or list of matrices to be converted to triplet form.
For \code{\link[=to_named]{to_named()}}, a data frame or list of data frames in triplet form to be converted to named form.}

\item{index_map}{A mapping between row and column names
and row and column indices.
See details.}

\item{row_index_colname, col_index_colname}{The names of row and column index columns in data frames.
Defaults are "i" and "j", respectively.}

\item{val_colname}{The name of the value column in data frames.
Default is "x".}

\item{rownames_colname, colnames_colname}{The name of row name and column name columns in data frames.
Defaults are "rownames" and "colnames", respectively.}

\item{matrix_class}{One of "matrix" (standard) or "Matrix" (sparse) representation
for matrices.
Default is "matrix".}
}
\value{
\code{\link[=to_triplet]{to_triplet()}} returns \code{a} as a data frame or list of data frames in triplet form.
\code{\link[=to_named]{to_named()}} returns \code{a} as a matrix or a list of matrices in named form.
}
\description{
Matrices can be in named form or indexed form.
Named form is the usual representation for the \code{matsindf} package,
wherein names for rows and columns are included in the \code{dimnames}
attribute of every matrix object, consuming memory.
Typically, neither zero rows nor zero columns are present.
In some instances,
many sparse matrices with the same names will be created,
leading to inefficiencies.
It can be more memory-efficient to store the matrices in
triplet form,
(a table format with matrix data represented as
a data frame with
row integer (i),
column integer (j), and
value (x) columns.
In triplet form,
a separate (external) mapping between
row and column indices and row and column names
must be maintained.
(In triplet form, it becomes the responsibility of the caller
to maintain a consistent mapping between row and column indices
and row and column names.
However, rowtype and coltype are retained as attributes
of the triplet data frame.)
These functions convert from named form to triplet form
(\code{\link[=to_triplet]{to_triplet()}})
and vice versa (\code{\link[=to_named]{to_named()}})
using externally supplied mappings
in the \code{index_map} argument.
\code{\link[=to_triplet]{to_triplet()}} and \code{\link[=to_named]{to_named()}} are exact inverses of each other.
}
\details{
\code{index_map} must be
an unnamed list of two data frames or
a named list of two or more data frames.
\itemize{
\item If an unnamed list of two data frames,
each data frame must have only
an integer column and a character column.
The first data frame of \code{index_map}
is interpreted as the mapping
between row names and row indices
and
the second data frame of \code{index_map}
is interpreted as the mapping
between column names and column indices.
\item If a named list of two or more data frames,
the names of \code{index_map}
are interpreted as row and column types,
with named data frames applied as the mapping for the
associated row or column type.
For example the data frame named "Industry" would be applied
to the dimension (row or column)
with an "Industry" type.
When both row and column have "Industry" type,
the "Industry" mapping is applied to both.
When sending named data frames in \code{index_map},
\code{a} must have both a row type and a column type.
If an appropriate mapping cannot be found in \code{index_map},
an error is raised.
Both matching data frames must have only
an integer column and
a character column.
}

When converting to indexed form,
rowtype and coltype
are retained as attributes.
See \code{\link[=rowtype]{rowtype()}} and \code{\link[=coltype]{coltype()}}.

If any indices are unavailable in the \code{index_map},
an error is raised.
It is an error to repeat a name in the name column of an \code{index_map}
data frame.
It is an error to repeat an index in the index column
of an \code{index_map} data frame.
}
