% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{transpose_byname}
\alias{transpose_byname}
\title{Transpose a matrix by name}
\usage{
transpose_byname(m)
}
\arguments{
\item{m}{the matrix to be transposed}
}
\value{
the transposed matrix
}
\description{
Transpose a matrix by name
}
\examples{
library(magrittr)
m <- matrix(c(11,21,31,12,22,32), ncol = 2, dimnames = list(paste0("i", 1:3), paste0("c", 1:2))) \%>\%
  setrowtype("Industry") \%>\% setcoltype("Commodity")
transpose_byname(m)
transpose_byname(list(m,m))
}
