% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{unaryapply_byname}
\alias{unaryapply_byname}
\title{Apply a unary function byname}
\usage{
unaryapply_byname(FUN, a)
}
\arguments{
\item{FUN}{a unary function to be applied "byname" to \code{a}.}

\item{a}{the argument to \code{FUN}.}
}
\value{
the result of applying \code{FUN} "byname" to \code{a}.
}
\description{
Apply a unary function byname
}
\examples{
library(magrittr)
productnames <- c("p1", "p2")
industrynames <- c("i1", "i2")
U <- matrix(1:4, ncol = 2, dimnames = list(productnames, industrynames)) \%>\%
  setrowtype("Products") \%>\% setcoltype("Industries")
difference_byname(0, U)
unaryapply_byname(`-`, U)
}
