% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{aggregate_pieces_byname}
\alias{aggregate_pieces_byname}
\title{Aggregate a matrix by pieces of row and/or column names}
\usage{
aggregate_pieces_byname(a, aggregation_map = NULL, piece, margin = c(1, 2))
}
\arguments{
\item{a}{}

\item{aggregation_map}{}

\item{piece}{}

\item{margin}{}
}
\value{
A version of \code{a} with rows and/or columns aggregated according to \code{aggregation_map}.
}
\description{
Aggregate a matrix (or list of matrices or a column in a \code{matsindf} data frame)
by pieces of the row and column names.
}
