% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply.R
\name{apply_byname}
\alias{apply_byname}
\title{Apply a function "by name" to any number of operands}
\usage{
apply_byname(FUN, ..., .FUNdots = NULL, match_type = c("all", "matmult",
  "none"), set_rowcoltypes = TRUE, .organize = TRUE)
}
\arguments{
\item{FUN}{a binary function to be applied "by name" sequentially to \code{...}.}

\item{...}{the operands for \code{FUN}.}

\item{.FUNdots}{a list of additional named arguments passed to \code{FUN}.}

\item{match_type}{one of "\code{all}", "\code{matmult}", or "\code{none}".
When both \code{a} and \code{b} are matrices,
"\code{all}" (the default) indicates that
rowtypes of \code{a} must match rowtypes of \code{b} and
coltypes of \code{a} must match coltypes of \code{b}.
If "\code{matmult}",
coltypes of \code{a} must match rowtypes of \code{b}.
If "\code{none}",
neither coltypes nor rowtypes are checked.}

\item{set_rowcoltypes}{tells whether to apply row and column types from \code{a} and \code{b}
to the output. 
Set \code{TRUE} (the default) to apply row and column types to the output.
Set \code{FALSE}, to \emph{not} apply row and column types to the output.}

\item{.organize}{a boolean that tells whether or not to automatically 
complete \code{a} and \code{b} relative to each other and
sort the rows and columns of the completed matrices.
Normally, this should be \code{TRUE} (the default).
However, if \code{FUN} takes over this responsibility, set to \code{FALSE}.}
}
\value{
the result of applying \code{FUN} sequantially to each item in \code{...}
}
\description{
\code{FUN} must be a binary function.
Arguments \code{match_type}, \code{set_rowcoltypes}, and \code{.organize}
have same meaning as on \code{\link[matsbyname]{binaryapply_byname}}.
}
\details{
\code{\link{apply_byname}} and \code{\link{cumapply_byname}} are similar.
Their differences can be described by considering a data frame.
\code{\link{apply_byname}} applies \code{FUN} to several columns (variables) of the data frame.
For example, \code{\link{sum_byname}} applied to several variables gives another column
containing the sums across each row of the data frame.
\code{\link{cumapply_byname}} applies \code{FUN} to successive entries in a single column.
For example \code{\link{sum_byname}} applied to a single column gives the sum of all numbers in that column.
}
