\name{byname-package}
\alias{byname-package}
\alias{byname}
\docType{package}
\title{
\packageTitle{byname}
}
\description{
\packageDescription{byname}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{byname}
\packageIndices{byname}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{byname}

Maintainer: \packageMaintainer{byname}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package 
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
