% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{coltype}
\alias{coltype}
\title{Column type}
\usage{
coltype(a)
}
\arguments{
\item{a}{The object from which you want to extract column types.}
}
\value{
The column type of \code{a}.
}
\description{
Extracts column type of \code{a}.
}
\examples{
commoditynames <- c("c1", "c2")
industrynames <- c("i1", "i2")
U <- matrix(1:4, ncol = 2, dimnames = list(commoditynames, industrynames)) \%>\%
  setrowtype(rowtype = "Commodities") \%>\% setcoltype("Industries")
coltype(U)
# This also works for lists
coltype(list(U,U))
}
