% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{constant_vector_byname}
\alias{constant_vector_byname}
\title{Creates a "byname" vector filled with a constant value k}
\usage{
constant_vector_byname(data, k, col_name, rowtype = NULL, coltype = NULL)
}
\arguments{
\item{data}{A matrix, or a column of a data frame populated with matrices, from which the "byname" i vector will be created.}

\item{k}{The constant that must be filled in for each coefficient of the created vector.}

\item{col_name}{A string containing the name of the i vector column.}

\item{rowtype}{A string containing the rowtype of the output vector.
Default is NULL.}

\item{coltype}{A string containing the coltype of the output vector.
Default is NULL.}
}
\value{
A "byname" vector, filled with constant values with the same number of rows, and rownames,
than the matrix that is provided as input.
}
\description{
Creates a "byname" constant vector (i.e. vector filled with a constant k value)
with the same number of rows and rownames than the matrix provided as input.
}
