% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{constant_vector_byname}
\alias{constant_vector_byname}
\title{Creates a "byname" vector filled with a constant value k}
\usage{
constant_vector_byname(
  data,
  k,
  col_name,
  rowtype = NULL,
  coltype = NULL,
  keep_rowcoltypes = c("all", "transpose", "row", "col", "none")
)
}
\arguments{
\item{data}{A matrix, or a column of a data frame populated with matrices, from which the "byname" i vector will be created.}

\item{k}{The constant that must be filled in for each coefficient of the created vector.}

\item{col_name}{A string containing the name of the i vector column.}

\item{rowtype}{A string containing the rowtype of the output vector.
Default is NULL.}

\item{coltype}{A string containing the coltype of the output vector.
Default is NULL.}

\item{keep_rowcoltypes}{A character string argument that defines whether column and rowtypes
should be kept from the matrix provided in data argument or not. See details.
Default is "all".}
}
\value{
A "byname" vector, filled with constant values with the same number of rows, and rownames,
than the matrix that is provided as input.
}
\description{
Creates a "byname" constant vector (i.e. vector filled with a constant \code{k} value)
with the same number of rows and rownames than the matrix provided as input.
}
\details{
Options for the keep_rowcoltypes argument are:
\itemize{
\item "all": transfer both row and column types of a directly to output.
\item "transpose": rowtype of a becomes coltype of output; coltype of a becomes rowtype of output.
"transpose" is helpful for FUNs that transpose a upon output.
\item "row": rowtype of a becomes both rowtype and coltype of output.
\item "col": coltype of a becomes both rowtype and coltype of output.
\item "none": owtype and coltype not set by unaryapply_byname. Rather, FUN will set rowtype and coltype.
}
}
\examples{
single_mat_2 <- create_matrix_byname(data = c(1, 2), nrow = 2, ncol = 1,
                                     dimnames = list(c("r1", "r2"), "c1"))
constant_vector_byname(single_mat_2, k = 8, col_name = "output_column")
list_of_mats <- create_matrix_byname(data = list(1, 2), nrow = list(1, 1), ncol = list(1,1), 
                                   dimnames = list(list("r1", "c1"), list("R1", "C1")))
constant_vector_byname(data = list_of_mats, k = 12,
                                  col_name = "output_column")
}
