% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{create_matrix_byname}
\alias{create_matrix_byname}
\title{Creates a "byname" matrix}
\usage{
create_matrix_byname(
  .dat,
  nrow,
  ncol,
  byrow = FALSE,
  dimnames,
  rowtype = NA,
  coltype = NA,
  rowcoltypes = c("all", "transpose", "row", "col", "none")
)
}
\arguments{
\item{nrow}{The number of rows to be used to create the matrix, in a list format, or as a data frame column
containing a list of the number of rows to be used for each observation.}

\item{ncol}{The number of columns to be used to create the matrix, in a list format, or as a data frame column
containing a list of the number of columns to be used for each observation.}

\item{byrow}{The argument stating whether the matrix should be filled by rows or by columns (FALSE by column, TRUE by row),
in a list format, or as a data frame column containing a list of the byrow argument for each observation.
Default is FALSE.}

\item{dimnames}{The dimension names to be used for creating the matrices, in a list format, or as a data frame column
containing a list of the dimension names to be used for each observation.}

\item{rowtype}{The row types to be used for creating the matrices, in a list format, or as a data frame column name
containing a list of the row types to be used for each observation.}

\item{coltype}{The column types to be used for creating the matrices, in a list format, or as a data frame column name
containing a list of the column types to be used for each observation.}

\item{data}{The data to be used to create the matrix, in a list format, or as a data frame column
containing a list of the data to be used for each observation.}

\item{keep_rowcoltypes}{A character string defining whether the row and column types (if available) of the input data (\code{.dat}) should be kept.
Options are "all", "none", "row", "col".
To use the rowtype and coltype arguments, the value should be set to "none".
Default is "all".}
}
\value{
A matrix, list of matrices, or column in a data frame, depending on the input arguments.
}
\description{
The function creates a "byname" matrix, or list of matrices, from \code{.dat}, depending on the input arguments.
}
\details{
If \code{rowtype} or \code{coltype} are \code{NA} (the default),
row and column types are taken from \code{.dat} and the \code{rowcoltypes} argument applies.
If either \code{rowtype} or \code{coltype} are set different from default (i.e., different from \code{NA}),
the values of the \code{rowtype} and \code{coltype} arguments override any row type or column type already present on \code{.dat}.
}
\examples{
create_matrix_byname(c(1, 2), nrow = 2, ncol = 1,
                     dimnames = list(c("r1", "r2"), "c1"))
create_matrix_byname(list(1, 2), nrow = list(1, 1), ncol = list(1,1), 
                     dimnames = list(list("r1", "c1"), list("R1", "C1")))
}
