% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{create_matrix_byname}
\alias{create_matrix_byname}
\title{Title}
\usage{
create_matrix_byname(
  data,
  nrow,
  ncol,
  byrow = FALSE,
  dimnames,
  rowtype = NULL,
  coltype = NULL,
  keep_rowcoltypes = c("all", "transpose", "row", "col", "none")
)
}
\arguments{
\item{data}{}

\item{nrow}{}

\item{ncol}{}

\item{byrow}{}

\item{dimnames}{}

\item{rowtype}{}

\item{coltype}{}
}
\value{

}
\description{
Title
}
