% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{create_rowvec_byname}
\alias{create_rowvec_byname}
\title{Create row vectors from data}
\usage{
create_rowvec_byname(
  .dat,
  rowname,
  rowcoltypes = c("all", "transpose", "row", "col", "none")
)
}
\arguments{
\item{.dat}{Data to be converted to row vectors.}

\item{rowname}{The name of the row of the rowvector.}

\item{rowcoltypes}{Tells how to handle row and column types already present in \code{.dat}.
Default is to keep both row type and column type.}
}
\value{
A row vector, a list of row vectors, or a data frame column of row vectors, depending on the
value of \code{.dat}.
}
\description{
This function takes data in the \code{.dat} and creates row vectors.
}
\details{
The column names in the resulting row vector are take from the names of \code{.dat}.

This function is a "byname" function that can accept a single number,
a vector, a list, or a data frame in \code{.dat}.

Row types and column types are taken from the row type and column type attributes of \code{.dat}.
}
\examples{
# Works with single numbers
create_rowvec_byname(c(c1 = 1) \%>\% setrowtype("rt") \%>\% setcoltype("ct"), 
                     rowname = "r1")
# Works with vectors
create_rowvec_byname(c(1, 2), dimnames = list("r1", c("c1", "c2")))
# Works with a list
create_rowvec_byname(list(c(1,2), c(3,4,5)), 
                     dimnames = list(list("r1", c("c1", "c2")), 
                                     list("R1", c("C1", "C2", "C3"))))
# Works in a data frame
df1 <- data.frame(dat = I(list()), dimnms = I(list()))
df1[[1, "dat"]] <- 1
df1[[2, "dat"]] <- c(2,3)
df1[[3, "dat"]] <- c(1,2,3,4,5,6)
df1[[1, "dimnms"]] <- list("r1", "c1")
df1[[2, "dimnms"]] <- list("R1", c("C1", "C2"))
df1[[3, "dimnms"]] <- list("r1", c("c1", "c2", "c3", "c4", "c5", "c6"))
df1
df1 <- df1 \%>\%
  dplyr::mutate(
    rowvec_col = create_rowvec_byname(dat, dimnames = dimnms)
  )
df1$rowvec_col[[1]]
df1$rowvec_col[[2]]
df1$rowvec_col[[3]]
}
