% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{create_rowvec_byname}
\alias{create_rowvec_byname}
\title{Title}
\usage{
create_rowvec_byname(
  .dat,
  dimnames,
  rowtype = NULL,
  coltype = NULL,
  keep_rowcoltypes = c("all", "transpose", "row", "col", "none")
)
}
\arguments{
\item{keep_rowcoltypes}{}

\item{margin}{Tells which dimension is long: \code{1} for a row vector, \code{2} for a column vector}
}
\value{

}
\description{
Title
}
