% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{elementexp_byname}
\alias{elementexp_byname}
\title{Exponential of matrix elements}
\usage{
elementexp_byname(M)
}
\arguments{
\item{M}{a matrix of list of matrices}
}
\value{
M with each element replaced by its exponential
}
\description{
Gives the exponential of all elements of a matrix or list of matrices
}
\examples{
library(magrittr)
elementexp_byname(1)
m <- matrix(c(log(10),log(1),log(1),log(100)), 
  nrow = 2, dimnames = list(paste0("i", 1:2), paste0("c", 1:2))) \%>\%
  setrowtype("Industry") \%>\% setcoltype("Commodity")
elementexp_byname(m)
}
