% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary.R
\name{equal_byname}
\alias{equal_byname}
\title{Compare two matrices "by name"}
\usage{
equal_byname(a, b)
}
\arguments{
\item{a}{the first matrix to be compared}

\item{b}{the second matrix to be compared}
}
\value{
\code{TRUE} iff row and column types are same \emph{and}
row and column names are same \emph{and}
all entries in the matrix are same.
}
\description{
Matrices are completed and sorted relative to one another before comparison
}
\examples{
library(magrittr)
a <- matrix(1:4, nrow = 2)
b <- matrix(1:4, nrow = 2)
equal_byname(a, b)
a <- a \%>\% setrowtype("Industries") \%>\% setcoltype("Commodities")
equal_byname(a, b) # FALSE because a has row and column types, but b does not.
b <- b \%>\% setrowtype("Industries") \%>\% setcoltype("Commodities")
equal_byname(a, b)
dimnames(a) <- list(c("i1", "i2"), c("c1", "c2"))
dimnames(b) <- list(c("c1", "c2"), c("i1", "i2"))
equal_byname(a, b) # FALSE, because row and column names are not equal
dimnames(b) <- dimnames(a)
equal_byname(a, b)
}
