% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{hatize_byname}
\alias{hatize_byname}
\title{Creates a diagonal "hat" matrix from a vector.}
\usage{
hatize_byname(v)
}
\arguments{
\item{v}{The vector from which a "hat" matrix is to be created.

A "hat" matrix is one in which the only non-zero elements are stored on the diagonal.
To "hatize" a vector is to place its elements on the diagonal of an otherwise-zero square matrix.
\code{v} must be a matrix with at least one dimension of length 1 (a vector).
The names of both dimensions of the hatized matrix are the same and taken from \code{v}.
Note that the vector names are sorted prior to forming the "hat" matrix.}
}
\value{
A square "hat" matrix with size equal to the length of \code{v}.
}
\description{
Creates a diagonal "hat" matrix from a vector.
}
\examples{
v <- matrix(1:10, ncol = 1, dimnames=list(c(paste0("i", 1:10)), c("c1"))) \%>\%
  setrowtype("Industries") \%>\% setcoltype(NA)
r <- matrix(1:5, nrow = 1, dimnames=list(c("r1"), c(paste0("c", 1:5)))) \%>\%
  setrowtype(NA) \%>\% setcoltype("Commodities")
hatize_byname(v)
hatize_byname(r)
# This also works with lists.
hatize_byname(list(v, v))
}
