% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{i_byname}
\alias{i_byname}
\title{Title}
\usage{
i_byname(a, col_name)
}
\arguments{
\item{a}{A matrix or list of matrices from which row or column names}

\item{col_name}{}

\item{row_names}{}
}
\value{

}
\description{
Title
}
