% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{i_byname}
\alias{i_byname}
\title{Title}
\usage{
i_byname(data, col_name, rowtype = NULL, coltype = NULL)
}
\arguments{
\item{data}{}

\item{col_name}{}

\item{rowtype}{}

\item{coltype}{}
}
\value{

}
\description{
Title
}
