% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{identize_byname}
\alias{identize_byname}
\title{Named identity matrix}
\usage{
identize_byname(m)
}
\arguments{
\item{m}{The matrix whose names and dimensions are to be preserved in an identity matrix.}
}
\value{
A square identity matrix with ones on the diagonal. 
Row and column names are taken from row and column
names of \code{m}.
Row and column names are sorted.
}
\description{
Creates an identity matrix (I) with ones on the diagonal with same size and names as \code{m}.
\code{m} must be square.
}
\examples{
m <- matrix(1:16, ncol = 4, dimnames=list(c(paste0("i", 1:4)), paste0("c", 1:4))) \%>\%
  setrowtype("Industries") \%>\% setcoltype("Commodities")
identize_byname(m)
n <- matrix(c(-21, -12, -21, -10), ncol = 2, dimnames = list(c("b", "a"), c("b", "a"))) \%>\%
  setrowtype("Industries") \%>\% setcoltype("Commodities")
identize_byname(n)
# This also works with lists
identize_byname(list(m, m))
}
