% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{list_of_rows_or_cols}
\alias{list_of_rows_or_cols}
\title{Named list of rows or columns of matrices}
\usage{
list_of_rows_or_cols(m, margin)
}
\arguments{
\item{m}{a matrix or list of matrices (say, from a column of a data frame)}

\item{margin}{the dimension of the matrices to be extracted (1 = rows, 2 = columns)}
}
\value{
a named list of rows or columns extracted from \code{m}
}
\description{
Extraction gives column vectors, regardless of margin.
}
\examples{
m <- matrix(data = c(1:6), nrow = 2, ncol = 3, dimnames = list(c("p1", "p2"), c("i1", "i2", "i3"))) \%>\%
setrowtype(rowtype = "Products") \%>\% setcoltype(coltype = "Industries")
list_of_rows_or_cols(m, margin = 1)
list_of_rows_or_cols(m, margin = 2)
}
