% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3matbyname.R
\name{mat_byname}
\alias{mat_byname}
\title{Create a \code{mat_byname} object}
\usage{
mat_byname(m)
}
\arguments{
\item{m}{a \code{matrix} object}
}
\value{
a \code{mat_byname} object
}
\description{
\code{m} should be a matrix.
}
\details{
If \code{is.null(m)}, \code{NULL} is returned.
If \code{all(is.na(m))}, \code{NA} is returned.
If \code{m} is not a \code{matrix}, an error is given.
}
\examples{
mat_byname(matrix(c(1:2)))
}
