% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{matricize_byname}
\alias{matricize_byname}
\title{Matricize a vector}
\usage{
matricize_byname(a, notation)
}
\arguments{
\item{a}{A row (column) vector to be converted to a matrix based on its row (column) names.}

\item{notation}{A string vector created by \code{RCLabels::notation_vec()} that identifies the notation for row or column names.}
}
\value{
A matrix created from vector \code{a}.
}
\description{
Converts a vector with rows or columns named according to \code{notation}
into a \code{matrix} or a \code{Matrix}, depending on the type of \code{a}.
Row and column types of the output are taken from the
row or column type of the long dimension of the incoming vector.
If the row or column type of the long dimension of the incoming vector is \code{NULL},
the outgoing matrix will have \code{NULL} rowtype and \code{NULL} coltype.
}
\examples{
v <- matrix(c(1,
              2,
              3, 
              4), 
            nrow = 4, ncol = 1, dimnames = list(c("p1 -> i1", 
                                                  "p2 -> i1", 
                                                  "p1 -> i2", 
                                                  "p2 -> i2"))) \%>\% 
  setrowtype("Products -> Industries")
# Default separator is " -> ".
matricize_byname(v, notation = RCLabels::arrow_notation)
}
