% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{rename_via_pattern_byname}
\alias{rename_via_pattern_byname}
\title{Rename row or column names via regexp pattern}
\usage{
rename_via_pattern_byname(
  a,
  margin = list(c(1, 2)),
  regexp_pattern,
  replacement,
  pieces = "all",
  prepositions = RCLabels::prepositions_list,
  notation = RCLabels::bracket_notation,
  ...
)
}
\arguments{
\item{a}{}

\item{margin}{}

\item{replacement}{}

\item{prepositions}{}

\item{notation}{}

\item{pattern}{}

\item{inf_notation}{}

\item{choose_most_specific}{}
}
\description{
Rename row or column names via regexp pattern
}
