% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{retain_remove}
\alias{retain_remove}
\title{Decides which columns to retain and remove.}
\usage{
retain_remove(row_col_names)
}
\arguments{
\item{row_col_names}{names of rows or columns to be retained or removed.}
}
\value{
a list with two items: \code{retain_names} and \code{remove_names}.
}
\description{
This is a helper function for \code{select_rows_byname} and \code{select_cols_byname}.
Specify removal by a leading \code{-}.
Names without a leading \code{-} will be retained.
Any number of leading \code{-} will be ignored when deciding the names of columns to be removed.
}
\examples{
retain_remove("c1")
retain_remove("-c1")
retain_remove(c("c1", "-c1")) # Retain takes precedence.
names <- c("-c1", "-c4", "-----c3", "c2", "c4", "c4") # Multiple "-" ignored.
retain_remove(names)
}
