% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\name{row-col-notation}
\alias{row-col-notation}
\alias{row_col_notation_list}
\alias{arrow_notation_list}
\alias{paren_notation_list}
\alias{split_pref_suff}
\alias{join_pref_suff}
\alias{switch_pref_suff}
\alias{switch_row_col_notation}
\title{Row and column notation}
\usage{
row_col_notation_list(
  sep = " -> ",
  pref_start = NULL,
  pref_end = sep,
  suff_start = sep,
  suff_end = NULL
)

arrow_notation_list()

paren_notation_list(suff_start = " [", suff_end = "]")

split_pref_suff(x, notation_list = arrow_notation_list())

join_pref_suff(ps, notation_list = arrow_notation_list())

switch_pref_suff(x, notation_list = arrow_notation_list())

switch_row_col_notation(x, from, to, reverse = FALSE)
}
\arguments{
\item{sep}{a string separator between prefix and suffix. Default is " -> ".}

\item{pref_start}{a string indicating the start of a prefix. Default is \code{NULL}.}

\item{pref_end}{a string indicating the end of a prefix. Default is the value of \code{sep}.}

\item{suff_start}{a string indicating the start of a suffix. Default is the value of \code{sep}.}

\item{suff_end}{a string indicating the end of a suffix. Default is \code{NULL}.}

\item{x}{a string or list of strings to be operated upon}

\item{notation_list}{a notation list generated by one of the \verb{*_notation_list()} functions, such as
\code{row_col_notation_list()}, \code{arrow_notation_list()}, or \code{paren_notation_list()}.
Default is \code{arrow_notation_list()}.}

\item{ps}{a list of prefixes and suffixes in which each item of the list is itself a list with two items, \code{pref} and \code{suff}}

\item{from}{the \code{notation_list} to switch \emph{away from}}

\item{to}{the \code{notation_list} to switch \emph{to}}
}
\value{
a string list with named items \code{pref_start}, \code{pref_end}, \code{suff_start}, and \code{suff_end}.
}
\description{
It is often convenient to represent row and column names
with notation that includes a prefix and a suffix,
with corresponding separators or start-end string sequences.
These functions assist with manipulating row and column names of this type.

notation symbols in a standard format
that is used by \code{matsbyname} in several places.

There are several functions that call \code{row_col_notation_list()} to generate specialized versions.
\itemize{
\item \code{row_col_notation_list()} builds lists of notation symbols in a standard format
that is used by \code{matsbyname} in several places.
By default, it builds a list of notation symbols that provides an arrow
separator (" -> ") between prefix and suffix.
\item \code{arrow_notation()} builds a list of notation symbols that provides an arrow separator (" -> ")
between prefix and suffix.
\item \code{paren_notation()} builds a list of notation symbols that provides square brackets (" \link{suffix}") around the suffix.
\item \code{split_pref_suff()} splits prefixes from suffixes, returning each in a list with names \code{pref} and \code{suff}.
If no delimiters are found, \code{x} is returned in the \code{pref} item, unmodified.
\item \code{join_pref_suff()} joins prefixes and suffixes, the inverse of \code{split_pref_suff()}.
\item \code{switch_pref_suff()} switches the location of prefix and suffix, such that the prefix becomes the suffix, and
the suffix becomes the prefix.
E.g., "a -> b" becomes "b -> a" or "a [b]" becomes "b [a]".
}
}
\section{\code{switch_row_col_notation()} switches from one type of notation to another based on the \code{from} and \code{to} arguments.}{
If \code{sep} only is specified (default is " -> "),
\code{pref_start}, \code{pref_end}, \code{suff_start}, and \code{suff_end} are
set appropriately.

None of the strings in a notation list are considered part of the prefix or suffix.
E.g., "a -> b" in arrow notation means that "a" is the prefix and "b" is the suffix.
}

\examples{
row_col_notation_list()
arrow_notation_list()
paren_notation_list()

}
