% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary.R
\name{samestructure_byname}
\alias{samestructure_byname}
\title{Test whether two matrices or lists of matrices have same structure}
\usage{
samestructure_byname(...)
}
\arguments{
\item{a}{the first matrix to be tested}

\item{b}{the second matrix to be tested}
}
\value{
\code{TRUE} if \code{a} and \code{b} have the same structure, \code{FALSE} otherwise.
}
\description{
Two matrices are said to have the same structure 
if row and column types are identical 
and 
if row and column names are identical.
}
\examples{
library(magrittr)
samestructure_byname(2, 2)
productnames <- c("p1", "p2")
industrynames <- c("i1", "i2")
U <- matrix(1:4, ncol = 2, dimnames = list(productnames, industrynames)) \%>\%
  setrowtype("Products") \%>\% setcoltype("Industries")
samestructure_byname(U, U)
samestructure_byname(U, U \%>\% setrowtype("row"))
samestructure_byname(U \%>\% setcoltype("col"), U)
# Also works with lists
samestructure_byname(list(U, U))
}
