% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Byname.R
\name{setrowtype}
\alias{setrowtype}
\title{Sets row type for a matrix or a list of matrices}
\usage{
setrowtype(x, rowtype)
}
\arguments{
\item{x}{the matrix on which row type is to be set}

\item{rowtype}{the type of item stored in rows}
}
\value{
a copy of \code{x} with \code{rowtype} attribute set
}
\description{
This function is a wrapper for \code{attr} so setting can be accomplished by the chain operator (\code{\%>\%})
in a functional way.
Row types are strings stored in the \code{rowtype} attribute.
}
\examples{
commoditynames <- c("c1", "c2")
industrynames <- c("i1", "i2")
U <- matrix(1:4, ncol = 2, dimnames = list(commoditynames, industrynames))
U \%>\% setrowtype("Commodities")
# This also works for lists
setrowtype(list(U,U), rowtype = "Commodities")
setrowtype(list(U,U), rowtype = list("Commodities", "Commodities"))
DF <- data.frame(U = I(list()))
DF[[1,"U"]] <- U
DF[[2,"U"]] <- U
setrowtype(DF$U, "Commodities")
DF <- DF \%>\% mutate(newcol = setrowtype(U, "Commodities"))
DF$newcol[[1]]
DF$newcol[[2]]
}
