% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{transpose_byname}
\alias{transpose_byname}
\title{Transpose a matrix by name}
\usage{
transpose_byname(a)
}
\arguments{
\item{a}{The matrix to be transposed.}
}
\value{
The transposed matrix.
}
\description{
Gives the transpose of a matrix or list of matrices.
}
\examples{
m <- matrix(c(11,21,31,12,22,32), ncol = 2, dimnames = list(paste0("i", 1:3), paste0("c", 1:2))) \%>\%
  setrowtype("Industry") \%>\% setcoltype("Commodity")
m
transpose_byname(m)
transpose_byname(list(m,m))
}
