% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{vec_from_store_byname}
\alias{vec_from_store_byname}
\title{Create a vector}
\usage{
vec_from_store_byname(m, store_vec, piece, colname = NA, column = TRUE)
}
\arguments{
\item{m}{}

\item{store_vec}{}

\item{piece}{}

\item{colname}{}

\item{column}{}
}
\value{

}
\description{
Create a vector
}
