% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{vectorize_byname}
\alias{vectorize_byname}
\title{Vectorize a matrix}
\usage{
vectorize_byname(a, notation_list)
}
\arguments{
\item{a}{the matrix to be vectorized.}

\item{notation_list}{a string list created by \code{row_col_notation_list()}.}
}
\value{
a column vector containing all elements of \code{a}, with row names assigned as "rowname \code{sep} colname".
}
\description{
Converts a matrix into a column vector.
Each element of the matrix becomes an entry in the column vector,
with rows named as "rowname \code{sep} colname" of the matrix entry.
If "colname \code{sep} rowname" is desired,
transpose the matrix first with \code{\link[=transpose_byname]{transpose_byname()}}.
}
\details{
\code{rowtype} and \code{coltype} attributes are retained in the event that
the resulting vector is re-matricized with the \code{\link[=matricize_byname]{matricize_byname()}} function later.
}
\examples{
m <- matrix(c(1, 5,
              4, 5),
            nrow = 2, ncol = 2, byrow = TRUE, 
            dimnames = list(c("p1", "p2"), c("i1", "i2"))) \%>\% 
  setrowtype("Products") \%>\% setcoltype("Industries")
m
vectorize_byname(m, notation_list = arrow_notation_list())
# If a single number is provided, the number will be returned as a 1x1 column vector 
# with some additional attributes.
vectorize_byname(42, notation_list = arrow_notation_list())
attributes(vectorize_byname(42, notation_list = arrow_notation_list()))
}
