% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unary.R
\name{vectorize_byname}
\alias{vectorize_byname}
\title{Vectorize a matrix}
\usage{
vectorize_byname(a, sep = " ", dimattr = "matdims_byname")
}
\arguments{
\item{a}{the matrix to be vectorized}

\item{sep}{a string to separate row names and col names in the resulting column vector. Default is " " (a space).}

\item{matdims_byname}{the name of the attribute in which the original dimensions of `a` are stored. Default is "matdims_byname".}
}
\value{
a vector with all elements of the matrix, with row names assigned as "rowname `sep` colname".
}
\description{
Converts a matrix into a column vector.
Each element of the matrix becomes an entry in the column vector,
with rows named as "rowname `sep` colname" of the matrix entry.
}
\details{
`rowtype` and `coltype` attributes are retained in the event that 
the resulting vector is re-matricized with the `matricize_byname` function later.
The number of rows and columns of `a` are recorded in attribute `dims`.
}
